alteração na declaração de campos

+++ src/ConectaPorto/Models/Usuario.cs
@@ -22,7 +22,7 @@ namespace ConectaPorto.Models
         }
         public int Id { get; set; }
         public string Nome { get; set; }
-        public int CPF { get; set; }
+        public long CPF { get; set; }
         public string Login { get; set; }
         public string Password { get; set; }
         public string PasswordHash { get; set; }
